# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal AI Doctor Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: doctors
    description: Doctor profiles and availability
  - name: appointments
    description: Appointment booking and management
  - name: medications
    description: Medication tracking and reminders
  - name: recovery
    description: Recovery progress tracking
  - name: dashboard
    description: Summary and analytics
  - name: openai
    description: OpenAI AI chat, image, and audio operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List all doctors
      parameters:
        - name: specialty
          in: query
          required: false
          schema:
            type: string
        - name: available
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"

  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get a doctor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List all appointments
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /medications:
    get:
      operationId: listMedications
      tags: [medications]
      summary: List all medications
      responses:
        "200":
          description: List of medications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: createMedication
      tags: [medications]
      summary: Add a medication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationInput"
      responses:
        "201":
          description: Created medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"

  /medications/{id}:
    patch:
      operationId: updateMedication
      tags: [medications]
      summary: Update a medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationUpdate"
      responses:
        "200":
          description: Updated medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteMedication
      tags: [medications]
      summary: Delete a medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /recovery:
    get:
      operationId: listRecoveryEntries
      tags: [recovery]
      summary: List recovery entries
      responses:
        "200":
          description: List of recovery entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecoveryEntry"
    post:
      operationId: createRecoveryEntry
      tags: [recovery]
      summary: Log a recovery entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryEntryInput"
      responses:
        "201":
          description: Created recovery entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryEntry"

  /recovery/{id}:
    delete:
      operationId: deleteRecoveryEntry
      tags: [recovery]
      summary: Delete a recovery entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        experience:
          type: integer
        location:
          type: string
        bio:
          type: string
        available:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - specialty
        - rating
        - reviewCount
        - experience
        - location
        - bio
        - available
        - createdAt

    Appointment:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        doctorName:
          type: string
        doctorSpecialty:
          type: string
        scheduledAt:
          type: string
          format: date-time
        reason:
          type: string
        status:
          type: string
          enum: [upcoming, completed, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - doctorId
        - doctorName
        - doctorSpecialty
        - scheduledAt
        - reason
        - status
        - createdAt

    AppointmentInput:
      type: object
      properties:
        doctorId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        reason:
          type: string
        notes:
          type: string
      required:
        - doctorId
        - scheduledAt
        - reason

    AppointmentUpdate:
      type: object
      properties:
        scheduledAt:
          type: string
          format: date-time
        reason:
          type: string
        status:
          type: string
          enum: [upcoming, completed, cancelled]
        notes:
          type: string

    Medication:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        reminderTime:
          type: ["string", "null"]
        startDate:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
          format: date-time
        active:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - dosage
        - frequency
        - active
        - startDate
        - createdAt

    MedicationInput:
      type: object
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        reminderTime:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        notes:
          type: string
      required:
        - name
        - dosage
        - frequency

    MedicationUpdate:
      type: object
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        reminderTime:
          type: string
        active:
          type: boolean
        notes:
          type: string

    RecoveryEntry:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date-time
        feelingScore:
          type: integer
        symptoms:
          type: string
        notes:
          type: ["string", "null"]
        appointmentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - feelingScore
        - symptoms
        - createdAt

    RecoveryEntryInput:
      type: object
      properties:
        date:
          type: string
          format: date-time
        feelingScore:
          type: integer
          minimum: 1
          maximum: 10
        symptoms:
          type: string
        notes:
          type: string
        appointmentId:
          type: integer
      required:
        - feelingScore
        - symptoms

    DashboardSummary:
      type: object
      properties:
        upcomingAppointments:
          type: integer
        activeMedications:
          type: integer
        recoveryStreak:
          type: integer
        lastFeelingScore:
          type: ["integer", "null"]
        totalAppointments:
          type: integer
        recentRecoveryEntries:
          type: array
          items:
            $ref: "#/components/schemas/RecoveryEntry"
        nextAppointment:
          $ref: "#/components/schemas/Appointment"
      required:
        - upcomingAppointments
        - activeMedications
        - recoveryStreak
        - totalAppointments
        - recentRecoveryEntries

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error