# lib/api-zod/src/generated/types/appointmentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal AI Doctor Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentUpdateStatus } from './appointmentUpdateStatus';

export interface AppointmentUpdate {
  scheduledAt?: Date;
  reason?: string;
  status?: AppointmentUpdateStatus;
  notes?: string;
}