# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal AI Doctor Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from './appointment';
import type { RecoveryEntry } from './recoveryEntry';

export interface DashboardSummary {
  upcomingAppointments: number;
  activeMedications: number;
  recoveryStreak: number;
  /** @nullable */
  lastFeelingScore?: number | null;
  totalAppointments: number;
  recentRecoveryEntries: RecoveryEntry[];
  nextAppointment?: Appointment;
}