# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal AI Doctor Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  rating: number;
  reviewCount: number;
  experience: number;
  location: string;
  bio: string;
  available: boolean;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: Date;
}