# lib/api-zod/src/generated/types/medicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal AI Doctor Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface MedicationInput {
  name: string;
  dosage: string;
  frequency: string;
  reminderTime?: string;
  startDate?: Date;
  endDate?: Date;
  notes?: string;
}